springboot-junit5-test-2/src/main/java/com/ercan/controller/BookController.java
package com.ercan.controller;

import com.ercan.entity.Book;
import com.ercan.exception.BookNotFoundException;
import com.ercan.service.BookService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;
import java.util.List;

@RestController
@RequestMapping("/api/book")
public class BookController {

    @Autowired
    private BookService bookService;


    @GetMapping("/{id}")
    public ResponseEntity<Book> getBookById(@PathVariable Long id){
        return ResponseEntity.ok(bookService.getBookById(id));
    }

    @GetMapping
    public ResponseEntity<List<Book>> getAllBook(){
        return ResponseEntity.ok(bookService.getAllBook());
    }

    @PostMapping
    public ResponseEntity<Book> saveBook(@RequestBody @Valid Book book){
        return ResponseEntity.ok(bookService.save(book));
    }

    @PutMapping
    public ResponseEntity<Book> updateBook(@RequestBody @Valid Book book) throws BookNotFoundException {
        return ResponseEntity.ok(bookService.update(book));
    }

}
